sprint4

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMassageService.java
@@ -4,7 +4,9 @@ import com.sprint.mission.discodeit.dto.binary.BinaryContentCreateRequestDto;
 import com.sprint.mission.discodeit.dto.message.CreateMessageRequestDto;
 import com.sprint.mission.discodeit.dto.message.UpdateMessageRequestDto;
 import com.sprint.mission.discodeit.entity.Message;
+import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.MessageRepository;
+import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.Interface.BinaryContentService;
 import com.sprint.mission.discodeit.service.Interface.MessageService;
 import lombok.RequiredArgsConstructor;
@@ -23,11 +25,21 @@ public class BasicMassageService implements MessageService {
     @Autowired
     private final MessageRepository messageRepository;
     @Autowired
-    private BinaryContentService binaryContentService;
+    private final BinaryContentService binaryContentService;
+    @Autowired
+    private final UserRepository userRepository;
+    @Autowired
+    private final ChannelRepository channelRepository;
 
 
     @Override
-    public Message createMessage(CreateMessageRequestDto request) throws Exception {
+    public Message createMessage(CreateMessageRequestDto request)  {
+        if(!channelRepository.existsById(request.getChannelId())) {
+            throw new NoSuchElementException("Channel not found");
+        }
+        if(!userRepository.existsById(request.getAuthorId())){
+            throw new NoSuchElementException("User not found");
+        }
         Message message=new Message(request.getContent(),request.getChannelId(),request.getAuthorId());
         Message savedMessage=messageRepository.save(message);
         if(request.getAttachments()!=null){
@@ -61,9 +73,9 @@ public class BasicMassageService implements MessageService {
     }
 
     @Override
-    public Message updateMessage(UpdateMessageRequestDto request) {
-        Message message = messageRepository.getMessageById(request.getMessageId())
-                .orElseThrow(() -> new NoSuchElementException("Message with id " + request.getMessageId() + " not found"));
+    public Message updateMessage(UUID id,UpdateMessageRequestDto request) {
+        Message message = messageRepository.getMessageById(id)
+                .orElseThrow(() -> new NoSuchElementException("Message with id " + id + " not found"));
         message.update(request.getNewContent());
         return messageRepository.save(message);
     }
@@ -71,6 +83,9 @@ public class BasicMassageService implements MessageService {
 
     @Override
     public void deleteMessage(UUID id) {
+        if(!messageRepository.existsById(id)) {
+            throw new NoSuchElementException("Message with id " + id + " not fount");
+        }
         binaryContentService.deleteByMessageId(id);
         messageRepository.deleteMessage(id);
     }
